requirements.txt
pytest~=8.1.1
allure-pytest~=2.13.5
selene>=2.0.0rc9
python-dotenv~=1.0.1
requests~=2.31.0
pydantic~=2.7.0
